backend/requirements.txt
google-genai
imap-tools
requests
fastapi
uvicorn[standard]
pydantic
python-dotenv
PyMuPDF
python-multipart
Pillow
imap-tools
requests
httpx
email-validator

google-api-python-client
google-auth-oauthlib
google-auth-httplib2


sqlalchemy
psycopg2-binary 
alembic 


python-jose[cryptography]  
passlib[bcrypt]


pydantic-settings